Fix deletion positions in alignment_tidy_ref_alt
The position_bc260 reference coordinates for deletions in the alternate sequence were not being incremented and thus treated like insertions. This leads to the start and end positions of these deletions in the reference to be truncated to 1bp, which causes many to incorrectly drop out during the window cleaning step.